store_function_based/requirements.txt
django
django-debug-toolbar
mysqlclient
djangorestframework